requirements.txt
Django~=5.1.7
djangorestframework~=3.15.2
psycopg-binary~=3.2.6
python-decouple~=3.8
django-filter~=25.1
django-extensions~=3.2.3